alzheimer_project/requirements.txt
# NeuroXAI Advanced - Colab-ready Dependencies
# Core ML
tensorflow>=2.14.0
keras
numpy>=1.23.0
pandas>=1.5.0

# Explainability
shap>=0.42.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.18.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# ML Utilities
scikit-learn>=1.3.0
umap-learn>=0.5.4
scipy>=1.11.0

# Data Handling
pyarrow>=14.0.0
nibabel>=5.1.0

# Jupyter
jupyter
nbformat
ipywidgets>=8.1.0

# Optional: Report Generation
# reportlab>=4.0.0
